move to App\Models and current User Model
